+++ hps_solver/tile_solver.cpp
@@ -3,8 +3,8 @@
 #include <iostream>
 
 
-std::vector<uint8_t> TileSolver::solveTile(std::shared_ptr<TileHeader> header, uint8_t iterations) {
-    std::vector<uint8_t> tile_data;
+std::vector<uint16_t> TileSolver::solveTile(std::shared_ptr<TileHeader> header, uint16_t iterations) {
+    std::vector<uint16_t> tile_data;
 
     complex origin = header->getOrigin();
     complex size = {header->getSize(), header->getSize()};
@@ -13,7 +13,7 @@ std::vector<uint8_t> TileSolver::solveTile(std::shared_ptr<TileHeader> header, u
     for (int y_index = 0; y_index < Constants::TILE_HEIGHT; y_index++) {
         for (int x_index = 0; x_index < Constants::TILE_WIDTH; x_index++) {
             complex c = {stride.real * x_index + origin.real, stride.imag * y_index + origin.imag};
-            uint8_t solution = solvePixel(c, iterations);
+            uint16_t solution = solvePixel(c, iterations);
             tile_data.push_back(solution);
         }
     }
@@ -22,12 +22,12 @@ std::vector<uint8_t> TileSolver::solveTile(std::shared_ptr<TileHeader> header, u
 }
 
 
-uint8_t TileSolver::solvePixel(complex c, uint8_t iterations) {
+uint16_t TileSolver::solvePixel(complex c, uint16_t iterations) {
     complex z = c;
 
     complex cycle_z = c;
 
-    for (uint8_t i = 1; i < iterations - 1; i++) {
+    for (uint16_t i = 1; i < iterations - 1; i++) {
         mpf_class z_real_new = (z.real * z.real) - (z.imag * z.imag) + c.real;
         mpf_class z_imag_new = (z.imag * z.real * 2) + c.imag;
         z = {.real = z_real_new, .imag = z_imag_new};
